Real support for dirrmtry
/!\ package format modified, still supporting the previous manifest /!\
/!\ new packages won't work with old pkgng /!\

To be able to still accept previous packages in manifest dirs: is still parsed,
but new directories will now be stored as directories:

for already installed packages, all directories are concidered as try to rm.

for plist @dirrmtry keyword is a known keyword no need to expand them in
bsd.port.mk unlike pkg_install

database upgrade is automatic

+++ libpkg/pkg_delete.c
@@ -130,8 +130,13 @@ pkg_delete_dirs(struct pkgdb *db, struct pkg *pkg, int force)
 		if (nbpackage > 1)
 			continue;
 
-		if (rmdir(pkg_dir_path(dir)) == -1 && errno != ENOTEMPTY && force != 1)
-			pkg_emit_errno("rmdir", pkg_dir_path(dir));
+		if (pkg_dir_try(dir)) {
+			if (rmdir(pkg_dir_path(dir)) == -1 && errno != ENOTEMPTY && force != 1)
+				pkg_emit_errno("rmdir", pkg_dir_path(dir));
+		} else {
+			if (rmdir(pkg_dir_path(dir)) == -1 && force != 1)
+				pkg_emit_errno("rmdir", pkg_dir_path(dir));
+		}
 	}
 
 	return (EPKG_OK);
